requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
diffusers>=0.23.0
accelerate>=0.24.0
trimesh>=4.0.0
numpy>=1.24.0
Pillow>=10.0.0
xformers>=0.0.22
fastapi>=0.104.0
uvicorn>=0.24.0
opencv-python>=4.8.0
scipy>=1.11.0
timm>=0.9.0
ddgs>=0.1.0
requests>=2.31.0
rembg>=2.0.50
gradio>=4.0.0
xatlas>=0.0.9
einops>=0.7.0
omegaconf>=2.3.0
realesrgan>=0.3.0
basicsr>=1.4.2
pyvista>=0.43.0
ollama>=0.1.6
chromadb>=0.4.22
sentence-transformers>=2.3.1
onnxruntime>=1.17.0
soundfile>=0.12.0
